make booklets ('pamphlets') without fancy commands
`pdfpages` seemed to be in my TeXlive installation by default, so it
might be less obnoxious for people to meet that requirement.

The LaTeX package `pdfpages` seems to do the same job as the many
commands involving PostScript in `make booklet`.
In the future, it might be elegant to run pandoc with a modified page
size before running it through `pdfpages`.