rewrote the Makefile
It now has easier to meet dependencies.

The PDF Toolkit is more annoying to get nowadays, since it depends on
deprecated Java stuff. (I'm Not sure if that's still true, but it was
for a few years.) So, I wanted to get rid of it (even though it's a very
nice suite of tools).

The pdfpages package in LaTeX is now what's used to collate the logical
pages for printing as a pamphlet/booklet, with the ability to specify
extra space for the fold.

bc and pdflatex have been added as dependencies, but they're extremely
reliable. Pandoc is still a dependency.

Things that might be nice in the future:
- a recipe/script to build the cover page, rather than making it
  manually in Inkscape every time
- recipies for team-specific booklets
- organizing the chapters so they don't spam up the top level dir